Add "assign to" custom relation
Is able to figure out if an assigned to / default assignee reference / relation is to a user or a group of users, and obtain that resource. It does sadly come at a cost of additional requests, yet there is no better solution that I can think of, for the time being.